feat: move fixtures into a separate fixtures folder
Previously, fixtures were placed in TestNameXXX_fixtures directories.
As the number of tests grows, this would clutter the test package with
many subfolders.

To avoid this, a dedicated fixtures folder has been created. Going
forward, all test fixtures for a package will be placed in this folder
to keep the root package clean and organized.