refactor(worker): standardize worker and scheduler startup pattern
Previously, the call to worker.start() was a non-blocking call,
because internally the async server startup was written inside
a go-routine. But the call to scheduler.start() was a blocking call.

This wasn't a consistent behaviour and it also meant if we changed
the order of worker startup and scheduler startup, our program will
not work as expected.

In this change, I have standardized the behaviour of both worker.start()
and scheduler.start(), making both of them as blocking operation. Hence,
the caller needs to start them as a separate go-routine. This also fixes
the problem of order of worker and scheduler startup. Now even if we
interchange the order, our program works as expected.