feat: add TaskScheduler interface to decouple from asynq.Scheduler
Previously, the scheduler code was directly dependent on
asynq.Scheduler, hence making it tightly coupled.

I have added a TaskScheduler interface which allows the code to
depend on the behavior of a scheduler rather than its implementation.

Hence, making the code more flexible and easier to replace with
alternative scheduler implementations if needed in the future.