feat: add helper function to group DB queries belonging to the same transaction in a parent span
In Bun, each DB query creates its own span, even when executed
inside a transaction. This makes it difficult to see which queries
belong to the same transaction

The `RunInTransaction` function addresses this by creating a parent
span with the provided `txName`. All queries executed within the
transaction become child spans of this parent, allowing them to be
visualized as a single unit of work in APM traces on Kibana